+++ Codigo-Fonte/API/ArpeFitness/src/main/java/pt/iade/ArpeFitness/models/repositories/SerieRepository.java
@@ -3,6 +3,5 @@ package pt.iade.ArpeFitness.models.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pt.iade.ArpeFitness.models.tables.Serie;
 
-public interface SerieRepository extends JpaRepository<Serie, Long> {
-    
+public interface SerieRepository extends JpaRepository<Serie, Integer> {
 }
